requirements.txt
flask-cors>=4.0
flask>=2.3
pdfplumber>=0.10
PyPDF2>=3.0
pandas>=2.0
openpyxl>=3.1
xlrd>=2.0
python-docx>=1.0
python-pptx>=0.6
pillow>=10.0
pytesseract>=0.3
opencv-python-headless>=4.8
numpy>=1.24
werkzeug>=3.0
gunicorn>=21.2
